+++ src/helpers.ts
@@ -66,6 +66,8 @@ import {TronSigner} from './tron/signer';
 import {CreateSmartContract} from './tron/types';
 
 let LedgerSigner: any; // TODO type
+let ethersprojectHardwareWalletsModule: any | undefined;
+let andrestEthersLedgerModule: any | undefined;
 let TrezorSigner: any; // TODO type
 let hardwareSigner: any; // TODO type
 
@@ -1704,16 +1706,18 @@ Note that in this case, the contract deployment will not behave the same if depl
               let error: any | undefined;
               try {
                 // eslint-disable-next-line @typescript-eslint/no-var-requires
-                const hardwareWalletModule = require('@ethersproject/hardware-wallets');
-                LedgerSigner = hardwareWalletModule.LedgerSigner;
+                ethersprojectHardwareWalletsModule = require('@ethersproject/hardware-wallets');
+                LedgerSigner = ethersprojectHardwareWalletsModule.LedgerSigner;
               } catch (e) {
                 error = e;
                 try {
                   // eslint-disable-next-line @typescript-eslint/no-var-requires
-                  const hardwareWalletModule = require('@anders-t/ethers-ledger');
-                  LedgerSigner = hardwareWalletModule.LedgerSigner;
+                  andrestEthersLedgerModule = require('@anders-t/ethers-ledger');
+                  LedgerSigner = andrestEthersLedgerModule.LedgerSigner;
                   error = undefined;
-                } catch (e) {}
+                } catch (e) {
+                  error = e;
+                }
               }
 
               if (error) {
@@ -1733,8 +1737,17 @@ Note that in this case, the contract deployment will not behave the same if depl
               ledgerSigner = undefined;
             }
 
-            ethersSigner = new LedgerSigner(provider);
-            hardwareWallet = 'ledger';
+            if (ethersprojectHardwareWalletsModule) {
+              derivationPath = getDerivationPath(network.config.chainId);
+              ethersSigner = new LedgerSigner(
+                provider,
+                'default',
+                derivationPath
+              );
+            } else {
+              ethersSigner = new LedgerSigner(provider, registeredProtocol);
+            }
+
             ledgerSigner = ethersSigner;
           } else if (registeredProtocol.startsWith('trezor')) {
             if (!TrezorSigner) {