+++ src/index.ts
@@ -12,6 +12,7 @@ import {
   NetworkConfig,
 } from 'hardhat/types';
 import {createProvider} from 'hardhat/internal/core/providers/construction'; // TODO harhdat argument types not from internal
+import {LazyInitializationProviderAdapter} from "hardhat/internal/core/providers/lazy-initialization";
 import {Deployment, ExtendedArtifact} from '../types';
 import {extendEnvironment, task, subtask, extendConfig} from 'hardhat/config';
 import {HARDHAT_NETWORK_NAME, HardhatPluginError} from 'hardhat/plugins';
@@ -307,6 +308,7 @@ extendEnvironment((env) => {
       );
     }
   }
+  initCompanionNetworks(env);
   log('ready');
 });
 
@@ -358,7 +360,7 @@ function setupExtraSolcSettings(settings: {
   // addIfNotPresent(settings.outputSelection["*"][""], "ast");
 }
 
-async function initCompanionNetworks(hre: HardhatRuntimeEnvironment) {
+function initCompanionNetworks(hre: HardhatRuntimeEnvironment) {
   hre.companionNetworks = {};
   for (const name of Object.keys(hre.network.companionNetworks)) {
     const networkName = hre.network.companionNetworks[name];
@@ -387,11 +389,13 @@ async function initCompanionNetworks(hre: HardhatRuntimeEnvironment) {
       throw new Error(`no network named ${networkName}`);
     }
 
-    network.provider = await createProvider(
-      hre.config,
-      networkName,
-      hre.artifacts
-    );
+    network.provider = new LazyInitializationProviderAdapter(() => {
+        return createProvider(
+          hre.config,
+          networkName,
+          hre.artifacts
+        );
+    })
 
     const networkDeploymentsManager = new DeploymentsManager(hre, network);
     deploymentsManager.addCompanionManager(name, networkDeploymentsManager);
@@ -452,7 +456,6 @@ subtask(TASK_DEPLOY_RUN_DEPLOY, 'deploy run only')
     if (typeof tags === 'string') {
       tags = tags.split(',');
     }
-    await initCompanionNetworks(hre);
     await deploymentsManager.runDeploy(tags, {
       log: args.log,
       resetMemory: false,