+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # hardhat-deploy
 
+## 0.11.45
+
+### Patch Changes
+
+- Init companion networks for all tasks (by tim-becker)
+
+## 0.11.44
+
+### Patch Changes
+
+- Ledger: add path support by @ifelsedeveloper
+
 ## 0.11.43
 
 ### Patch Changes